+++ core/src/com/mygdx/game/enemigos/Satelite.java
@@ -7,14 +7,13 @@ import com.mygdx.game.Hiriente;
 import com.mygdx.game.Nave;
 import com.mygdx.game.Util;
 
-public class Satelite extends Enemigo implements Hiriente{
-	private final static int ancho = 60;
-	private final static int alto = 40;
+public class Satelite extends Enemigo implements Hiriente {
+	private final static float ancho = 60;
+	private final static float alto = 40;
 	private final static int scoreChange = -20; // El satelite aun estaba en uso
 	private final static Texture image = new Texture(Gdx.files.internal("Satelite.png"));
 	
-	public Satelite(float x, float y, 
-            			float velX, float velY) {
+	public Satelite(float x, float y, float velX, float velY) {
 		super(x, y, ancho, alto, velX, velY, image);
 	}
